requirements.txt
pytest==6.1.2
azure-mgmt-resource==15.0.0
azure-cli-core==2.14.0
azure-mgmt-keyvault==8.0.0
azure-identity==1.4.1